# patchnotes-web/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "PatchNotes.Api | v1",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:5099/"
    }
  ],
  "paths": {
...
    "/api/users/me": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetCurrentUser",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "tags": [
          "Users"
        ],
        "operationId": "UpdateCurrentUser",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
    "/api/users/me/confirm-email-change": {
      "post": {
        "tags": [
          "Users"
        ],
        "operationId": "ConfirmEmailChange",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/me/email-preferences": {
      "get": {
        "tags": [
          "Users"
        ],
        "operationId": "GetEmailPreferences",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPreferencesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "tags": [
          "Users"
        ],
        "operationId": "UpdateEmailPreferences",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEmailPreferencesRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailPreferencesDto"
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "EmailPreferencesDto": {
        "type": "object",
        "properties": {
          "emailDigestEnabled": {
            "type": "boolean"
          },
          "emailWelcomeSent": {
            "type": "boolean"
          },
          "digestDay": {
            "type": "integer",
            "format": "int32"
          },
          "digestHour": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateEmailPreferencesRequest": {
        "required": [
          "emailDigestEnabled",
          "digestDay",
          "digestHour"
        ],
        "type": "object",
        "properties": {
          "emailDigestEnabled": {
            "type": [
              "null",
              "boolean"
            ]
          },
          "digestDay": {
            "type": "integer",
            "format": "int32"
          },
          "digestHour": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "UpdateUserRequest": {
        "required": [
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "null",
              "string"
            ]
          }
        }
      },
      "UserDto": {
        "required": [
          "id",
          "stytchUserId"
        ],
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "stytchUserId": {
            "type": "string"
          },
          "email": {
            "type": [
              "null",
              "string"
            ]
          },
          "name": {
            "type": [
              "null",
              "string"
            ]
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "lastLoginAt": {
            "type": [
              "null",
              "string"
            ],
            "format": "date-time"
          },
          "isPro": {
            "type": "boolean"
          },
          "isAdmin": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
}