# patchnotes-web/src/api/generated/model/updateUserRequest.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * PatchNotes.Api | v1
 * OpenAPI spec version: 1.0.0
 */

export interface UpdateUserRequest {
  /** @nullable */
  name: string | null;
}